Changes applied to version of DuckBurger we will be deploying
