--- contracts/src/example-plugins/DuckBurgerStateStorage/DuckBurgerHQStateStorage.yaml
@@ -1,17 +1,0 @@
-kind: BuildingKind
-spec:
-  name: Duck Burger HQ (State Storage)
-  description: "Play Ducks vs Burgers"
-  category: custom
-  model: 11-19
-  contract:
-    file: ./DuckBurgerHQStateStorage.sol
-  plugin:
-    file: ./DuckBurgerHQStateStorage.js
-  materials:
-    - name: Green Goo
-      quantity: 10
-    - name: Blue Goo
-      quantity: 10
-    - name: Red Goo
-      quantity: 10
